% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Say hello to someone as many times as you wish}
\usage{
hello(sayTo = "world", num = 1)
}
\arguments{
\item{sayTo}{- The name you want to say hello to.}

\item{num}{- Number to repeat hello.}
}
\value{
Prints hello statements to console
}
\description{
Say hello to someone as many times as you wish
}
